Adding the flags

+++ src/section.c
@@ -32,11 +32,11 @@ void	encrypt_section(char *file, Elf64_Shdr *header, char *key)
 static int	verif_sections(char *file, off_t file_size)
 {
 	off_t	headers_table;
-	int	headers_len;
+	int		headers_len;
 
 	headers_table = ((Elf64_Ehdr *)file)->e_shoff;
 	headers_len = ((Elf64_Ehdr *)file)->e_shnum;
-	if (headers_table + headers_len * sizeof(Elf64_Shdr) > file_size)
+	if (headers_table + headers_len * sizeof(Elf64_Shdr) > (size_t)file_size)
 	{
 		dprintf(2, "Corrupted binary\n");
 		return (0);
